27 12 21 gdz (#47)
* Amelioration de la fonction select_fabriques

* Deplacement des fonctions generer dans leur module et ajout d une option pour reprendre une sauvegarde

* Creation du bouton reprendre la partie

* Sauvegarde fonctionnelle 100 pourcents

* Bouton reprendre partie fonctionnel, il est present si le jeu detecte la presence du fichier save.txt

* Amelioration de la sauvegarde et ajout du score

* MAJ .gitignore

* Resolution du palais identique pour tous les joueurs au chargement d une sauvegarde

* Suppression de fichier inutiles

* gitingore

* Update save.txt

* Delete settings.txt

* Conditions de fin de partie et calcul du vainqueur, amelioration de la sauvegarde, resolutions de crashs,optimisations des performances

* Resolution du bug ou les bot ne pouvaient pas prendre de tuiles au milieu

+++ menu.py
@@ -29,9 +29,7 @@ def accueil():
     rectangle(100,340,300,390)
     texte(155,345,"Quitter")
 
-
     if existe('./files/save.txt'):
-
         rectangle(100,410,300,460)
         texte(110,415,"Reprendre partie",police='Arial',taille=17)
 
@@ -154,10 +152,10 @@ def clic_options(settings_config):
 
 def menu_jeu():
     fenetre()
-    config = copy_file("./files/settings.txt")
-    config[-1] = False
-    if config == []:
-        config = [2,[2],False,False]
+    config = [2,[2],False,False]
+    if existe("./files/settings.txt"):
+        config = copy_file("./files/settings.txt")
+        config[-1] = False
     print(config)
     
     while True:
@@ -198,3 +196,5 @@ def menu_jeu():
 
 if __name__ == '__main__':
     menu_jeu()
+=======
+
